Test github/requirements.txt
cmsplugin-filer==1.1.2
dj-database-url==0.4.1
Django==1.8.14
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.2.5
django-filer==1.2.4
django-formtools==1.0
django-mptt==0.8.6
django-polymorphic==0.8.1
django-reversion==1.8.7
django-sekizai==0.10.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.1.1
djangocms-column==1.6.0
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-installer==0.8.12
djangocms-link==1.7.2
djangocms-style==1.7.0
djangocms-text-ckeditor==2.9.5
easy-thumbnails==2.3
html5lib==0.9999999
Pillow==3.3.1
pytz==2016.6.1
six==1.10.0
tzlocal==1.2.2
Unidecode==0.4.19
yolk==0.4.3
yolk3k==0.9
